+++ examples/lexicons/codelingo/ast/typescript/src/ParserTypeScript.ts
